requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
google-genai>=0.2.0
simpleeval>=0.9.13
